Update T-Coffee download URL in README.md
The version we have previously used is now archived, rather than the latest stable release. The bug that requires us to set the MAX_N_PID environmental variable later has not been resolved in the current latest release, so for simplicity I suggest we use an archived release now - to avoid having to update the link later, except when the bug is fixed.