Correctly set MAX_N_PID_4_TCOFFEE in README.md
Sets the environment variable to the pid_max value provided by the OS kernel, instead of a hard-coded value. This is a workaround for a known bug in T-Coffee on 64-Bit Linux distributions (see Issue 27 on the T-Coffee GitHub repository). The bug itself was actually fixed four days ago on the development branch of T-Coffee (not yet available as a release), so this should not be necessary in a future release.